refactor(extracts use-subscription, merges api reference into tsdoc)
