+++ libs/react/src/lib/react.tsx
@@ -1,48 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Observable } from 'rxjs';
 
 export { useStore } from './use-store';
+export { useSubscription } from './use-subscription';
 export { store, StoreArg, StoreReturn } from './store';
-
-/**
- * A react hook for subscribing to an observable
- * it will subscribe to an observable given to it &
- * it will re-render your component anytime the observable
- * emits a value, produces an error, or completes.
- *
- * It will return a tuple of the latest value, the error
- * value, and the completion status (boolean):
- *
- * eg:
- *
- * subject.next(1)
- * subject.next(2)
- * subject.next(3)
- * subject.error('foo')
- *
- * In your component:
- *
- * const [count, error, complete] = useSubscription(obs$)
- * console.log({count, error, complete})
- *
- * Your component will render & log: 1, 2, 3, foo...
- */
-export function useSubscription<T>(
-  source: Observable<T>
-): [T | undefined, any, boolean] {
-  const [next, setNext] = useState<T | undefined>();
-  const [error, setError] = useState<any>();
-  const [complete, setComplete] = useState<boolean>(false);
-  useEffect(() => {
-    const subscription = source.subscribe(
-      (value) => setNext(value),
-      (error) => setError(error),
-      () => setComplete(true)
-    );
-    return () => subscription.unsubscribe();
-  }, [source]);
-  return [next, error, complete];
-}
 
 export function withSubscription<T>(
   WrappedComponent: React.ComponentType<{
